+++ app/components/FileDropdown.vue
@@ -12,6 +12,35 @@
             <template v-for="(option, index) in options" :key="option.name ?? index">
                 <DropdownMenuSeparator v-if="option.separator" />
 
+                <DropdownMenuLabel v-else-if="option.label" class="text-xs text-zinc-500">
+                    {{ option.title }}
+                </DropdownMenuLabel>
+
+                <DropdownMenuSub v-else-if="option.children">
+                    <DropdownMenuSubTrigger>
+                        {{ option.title }}
+                    </DropdownMenuSubTrigger>
+
+                    <DropdownMenuSubContent>
+                        <template
+                            v-for="(child, childIndex) in option.children"
+                            :key="child.name ?? childIndex"
+                        >
+                            <DropdownMenuSeparator v-if="child.separator" />
+
+                            <DropdownMenuItem v-else-if="child.href" as-child>
+                                <a :href="child.href" target="_blank">
+                                    {{ child.title }}
+                                </a>
+                            </DropdownMenuItem>
+
+                            <DropdownMenuItem v-else @select="child.click()">
+                                {{ child.title }}
+                            </DropdownMenuItem>
+                        </template>
+                    </DropdownMenuSubContent>
+                </DropdownMenuSub>
+
                 <DropdownMenuItem v-else-if="option.href" as-child>
                     <a :href="option.href" target="_blank">
                         {{ option.title }}